add simple stringstream

+++ MSTL/basiclib.cpp
@@ -2,7 +2,7 @@
 #include <assert.h>
 MSTL_BEGIN_NAMESPACE__
 
-void split_line(std::ostream& out, size_t size, char split_type) {
+void split_line(std::ostream& out, uint32_t size, char split_type) {
 	while (size--) out << split_type;
 	out << std::endl;
 }
